requirements.txt
requests==2.33.1
pandas==3.0.2
load-dotenv==0.1.0
gradio==6.12.0
arabic-reshaper==3.0.0
python-bidi==0.6.7
nltk==3.9.4
emoji==2.15.0
scikit-learn==1.8.0
farasapy==0.1.1
PyArabic==0.6.15
matplotlib==3.10.8
wordcloud==1.9.6
seaborn==0.13.2




